pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "produckai-mcp-server"
version = "0.7.0"
description = "ProduckAI MCP Server - Product feedback analysis integration for Claude Desktop"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "ProduckAI Team", email = "contact@produckai.com"}
]
keywords = ["mcp", "claude", "product-management", "feedback-analysis", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # MCP SDK
    "mcp>=1.0.0",

    # HTTP client
    "httpx>=0.27.0",
    "aiohttp>=3.9.0",

    # Data validation
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",

    # Configuration
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",

    # CLI
    "click>=8.1.0",
    "rich>=13.0.0",

    # OAuth & Security
    "keyring>=24.0.0",
    "cryptography>=41.0.0",

    # Integration SDKs
    "google-api-python-client>=2.100.0",
    "google-auth-httplib2>=0.1.1",
    "google-auth-oauthlib>=1.1.0",
    "slack-sdk>=3.26.0",
    "jira>=3.5.0",

    # Document Processing
    "PyPDF2>=3.0.0",

    # AI/LLM
    "anthropic>=0.39.0",
    "openai>=1.0.0",

    # Utilities
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
]

[project.scripts]
produckai-mcp = "produckai_mcp.cli:main"

[project.urls]
# TODO: Update these URLs when you create your GitHub repository
Homepage = "https://github.com/produckai/mcp-server"
Documentation = "https://github.com/produckai/mcp-server/blob/main/README.md"
Repository = "https://github.com/produckai/mcp-server"
Issues = "https://github.com/produckai/mcp-server/issues"
Changelog = "https://github.com/produckai/mcp-server/blob/main/CHANGELOG.md"

[tool.setuptools]
packages = ["produckai_mcp"]

[tool.setuptools.package-dir]
produckai_mcp = "src/produckai_mcp"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_calls = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "slack_sdk.*",
    "jira.*",
    "google.*",
    "googleapiclient.*",
    "PyPDF2.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src/produckai_mcp --cov-report=html --cov-report=term"
asyncio_mode = "auto"
